back/backend/requirements.txt
fastapi==0.135.3
SQLAlchemy==2.0.49
aiosqlite==0.22.1
uvicorn==0.44.0
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.4.0
pydantic==2.13.0
pydantic-settings==2.8.1
python-dotenv==1.2.2
python-multipart==0.0.26
reportlab==4.4.10
aiofiles==25.1.0
email-validator==2.3.0
